Moving ES-specific parts of the export backend into place.
