+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/views/export/ElasticsearchExportBackend.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.plugins.views.search.export.es;
+package org.graylog.storage.elasticsearch6.views.export;
 
 import com.google.inject.name.Named;
 import io.searchbox.core.Search;
